requirements.txt
requests==2.31.0
selenium==4.6.0
chromedriver-binary-auto==0.3.1
